integrate image creation.
turns out the manifest holds the blueprint for the image. so you use containerd to sort of map the manifest to an "image" and if you ever need to interact with that image, containerd will read the relevant blobs stored (using digests) and sort of create the necessary fs for you.